rust: make bindgen invalid Rust target detection conditional on <0.71
In 0.71.0 the bindgen output for invalid rust targets was changed to
"«version» is not a valid Rust target". This breaks the existing
detection.

Further investigation revealed that from 0.71.0 we probably don't
need to do this check; "all*" Rust versions will be accepted as valid
and converted internally to an appropriate target.

Bindgen does not like `-beta` or `-nightly` suffixes, we'll have
to address that separately depending on the outcome of
rust-lang/rust-bindgen#3152.

See-also: rust-lang/rust-bindgen#2993
Signed-off-by: Matt Jolly <kangie@gentoo.zip>